pyproject.toml
[project]
name = "vlog"
version = "0.1.0"
description = "VRChat Auto-Diary"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.26.0",
    "psutil>=5.9.0",
    "faster-whisper>=0.10.0",
    "google-generativeai>=0.8.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
